refactor: Saga 엔티티 및 enum을 application 레이어로 이동
