+++ order-service/src/main/java/com/palja/order_service/application/saga/model/OrderSagaStatus.java
@@ -1,4 +1,4 @@
-package com.palja.order_service.infrastructure.saga.model;
+package com.palja.order_service.application.saga.model;
 
 // DB 조회용
 public enum OrderSagaStatus {
