+++ order-service/src/main/java/com/palja/order_service/application/saga/steps/CreatePaymentStep.java
@@ -6,7 +6,7 @@ import com.palja.order_service.application.exception.OrderErrorCode;
 import com.palja.order_service.application.port.PaymentClient;
 import com.palja.order_service.application.saga.SagaStep;
 import com.palja.order_service.domain.entity.Order;
-import com.palja.order_service.infrastructure.saga.model.OrderSagaStep;
+import com.palja.order_service.application.saga.model.OrderSagaStep;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -48,17 +48,10 @@ public class CreatePaymentStep implements SagaStep {
             log.info("[SAGA][{}][SUCCESS] action=create, orderId={}, paymentId={}, elapsedMs={}",
                     name(), orderId, res.getPaymentId(), elapsedMs);
 
-        } catch (BusinessException e) {
-            long elapsedMs = (System.nanoTime() - start) / 1_000_000;
-            log.error("[SAGA][{}][FAIL] action=create, orderId={}, errorCode={}, elapsedMs={}, msg={}",
-                    name(), orderId, e.getErrorCode(), elapsedMs, e.getMessage(), e);
-            throw e;
-
         } catch (Exception e) {
             long elapsedMs = (System.nanoTime() - start) / 1_000_000;
             log.error("[SAGA][{}][ERROR] action=create, orderId={}, elapsedMs={}, msg={}",
                     name(), orderId, elapsedMs, e.getMessage(), e);
-//            throw e;
             throw new BusinessException(OrderErrorCode.PAYMENT_CREATION_FAILED);
         }
     }
