+++ order-service/src/main/java/com/palja/order_service/infrastructure/repository/impl/OrderSagaRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.palja.order_service.infrastructure.repository.impl;
 
-import com.palja.order_service.infrastructure.saga.model.OrderSaga;
+import com.palja.order_service.application.saga.model.OrderSaga;
 import com.palja.order_service.domain.repository.OrderSagaRepository;
 import com.palja.order_service.infrastructure.repository.JpaOrderSagaRepository;
 import lombok.RequiredArgsConstructor;
